Removed onSubmit handle from form onSubmit handle is in the Add Plant component.
